pyproject.toml
[project]
name = "rag2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "flashrank>=0.2.10",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-docling>=2.0.0",
    "langchain-openai>=1.1.6",
    "langsmith>=0.6.0",
    "openai>=2.14.0",
    "pypdf>=6.6.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "streamlit>=1.53.1",
]
